Rename constants.ts to overrides.ts, for clarity
